完成app demo基本框架。
